+++ src/main/kotlin/com/antlers/support/psi/impl/AntlersPhpBlockMixin.kt
@@ -1,0 +1,17 @@
+package com.antlers.support.psi.impl
+
+import com.intellij.extapi.psi.ASTWrapperPsiElement
+import com.intellij.lang.ASTNode
+import com.intellij.psi.LiteralTextEscaper
+import com.intellij.psi.PsiLanguageInjectionHost
+
+abstract class AntlersPhpBlockMixin(node: ASTNode) : ASTWrapperPsiElement(node), PsiLanguageInjectionHost {
+
+    override fun isValidHost(): Boolean = true
+
+    override fun updateText(text: String): PsiLanguageInjectionHost = this
+
+    override fun createLiteralTextEscaper(): LiteralTextEscaper<out PsiLanguageInjectionHost> {
+        return LiteralTextEscaper.createSimple(this)
+    }
+}